requirements.txt
yt-dlp>=2024.8.6
deepgram-sdk>=3.2.7
openai>=1.35.10
python-dotenv>=1.0.0
pydub>=0.25.1
ffmpeg-python>=0.2.0
requests>=2.32.2
fastapi>=0.111.0
uvicorn>=0.30.1
tqdm>=4.66.4